Update output of main driver function
Previously, the main driver function mhi(...) returned several different
objects, which changed depending on how the function was called.
The output has now been streamlined, with the function always returning
an instance of the class IrrepDecomposition. This class contains information
about the decomposition (irreps and matrices), the orbit (necessary for applying
the change of basis in practice), the orbit representation matrices (useful for
comparing to fiducial results), and the little/stabilizer groups.